split ltp task lists
