+++ user/src/bin/ltp_dependence/tasks/batches.rs
@@ -1,0 +1,855 @@
+//! Topic batches originally assembled as periodic "uncovered 100" pushes.
+//!
+//! Most arrays here are toggled live by the `RUN_*` booleans in
+//! [`super::super::submit_plan`]; a minority are still parked here as
+//! reference material (they ship the same shape as the active batches but
+//! their `RUN_*` flag has never been enabled). Names describe the syscall
+//! family covered; the historical `BATCH_YYYYMMDD_*` prefix has been
+//! dropped because it did not help with navigation.
+//
+// Auto-extracted from the original flat `mod.rs`; test names are
+// preserved verbatim so the set consumed by `submit_plan.rs` is unchanged.
+
+#![allow(dead_code)]
+
+// FS metadata + sync / stat / pipe / sendfile / splice / vmsplice group.
+// fs metadata+sync -> stat/statx -> pipe/sendfile/splice/vmsplice.
+const FALLOCATE_FSYNC_SYNC_TASKS: [&str; 20] = [
+    "fallocate01",
+    "fallocate02",
+    "fallocate03",
+    "fallocate04",
+    "fallocate05",
+    "fallocate06",
+    "fdatasync01",
+    "fdatasync02",
+    "fdatasync03",
+    "fsync01",
+    "fsync02",
+    "fsync03",
+    "fsync04",
+    "sync01",
+    "sync_file_range01",
+    "sync_file_range02",
+    "syncfs01",
+    "truncate03_64",
+    "readdir01",
+    "readdir21",
+];
+
+const STAT_LFS_EXT_TASKS: [&str; 20] = [
+    "fstat02_64",
+    "fstat03_64",
+    "fstatfs01",
+    "fstatfs01_64",
+    "fstatfs02_64",
+    "lstat01",
+    "lstat01_64",
+    "lstat02",
+    "lstat02_64",
+    "stat01_64",
+    "stat02_64",
+    "stat03_64",
+    "statfs01",
+    "statfs01_64",
+    "statfs02_64",
+    "statfs03_64",
+    "mknod07",
+    "mknodat02",
+    "statx04",
+    "statx05",
+];
+
+const STATX_EXT_TASKS: [&str; 7] = [
+    "statx06", "statx07", "statx08", "statx09", "statx10", "statx11", "statx12",
+];
+
+const PIPE_CORE_TASKS: [&str; 13] = [
+    "pipe01", "pipe02", "pipe03", "pipe04", "pipe05", "pipe06", "pipe07", "pipe08", "pipe09",
+    "pipe10", "pipe11", "pipe12", "pipe13",
+];
+
+const PIPE_SENDFILE_SPLICE_TASKS: [&str; 20] = [
+    "pipe14",
+    "pipe15",
+    "pipe2_01",
+    "pipe2_02",
+    "pipe2_04",
+    "sendfile02",
+    "sendfile02_64",
+    "sendfile03",
+    "sendfile03_64",
+    "sendfile04",
+    "sendfile04_64",
+    "sendfile05",
+    "sendfile05_64",
+    "sendfile06",
+    "sendfile06_64",
+    "splice01",
+    "splice02",
+    "splice03",
+    "splice04",
+    "splice05",
+];
+
+const TEE_VMSPLICE_FADVISE_TASKS: [&str; 20] = [
+    "sendfile07",
+    "sendfile07_64",
+    "sendfile08",
+    "sendfile08_64",
+    "sendfile09",
+    "sendfile09_64",
+    "splice06",
+    "splice07",
+    "splice08",
+    "splice09",
+    "tee01",
+    "tee02",
+    "vmsplice01",
+    "vmsplice02",
+    "vmsplice03",
+    "vmsplice04",
+    "posix_fadvise01",
+    "posix_fadvise02",
+    "posix_fadvise03",
+    "posix_fadvise04",
+];
+
+// Credential *_16 compatibility + POSIX MQ / SysV IPC core group.
+// Priority from ltp summary Section 2 + Section 8:
+// credential *_16 compatibility -> POSIX MQ + SysV IPC core.
+const CAP_CRED16_QUERY_TASKS: [&str; 20] = [
+    "cap_bounds_r",
+    "cap_bounds_rw",
+    "cap_bset_inh_bounds",
+    "capset03",
+    "capset04",
+    "check_pe",
+    "check_simple_capset",
+    "getegid01_16",
+    "getegid02_16",
+    "geteuid01_16",
+    "geteuid02_16",
+    "getgid01_16",
+    "getgid03_16",
+    "getgroups01_16",
+    "getgroups03",
+    "getgroups03_16",
+    "getresgid01_16",
+    "getresgid02_16",
+    "getresgid03_16",
+    "getresuid01_16",
+];
+
+const CRED16_MUTATION_TASKS: [&str; 40] = [
+    "getresuid02_16",
+    "getresuid03_16",
+    "getuid01_16",
+    "getuid03_16",
+    "setfsgid01_16",
+    "setfsgid02_16",
+    "setfsgid03",
+    "setfsgid03_16",
+    "setfsuid01_16",
+    "setfsuid02_16",
+    "setfsuid03_16",
+    "setfsuid04_16",
+    "setgid01_16",
+    "setgid02_16",
+    "setgid03_16",
+    "setgroups01_16",
+    "setgroups02_16",
+    "setgroups03",
+    "setgroups03_16",
+    "setgroups04",
+    "setgroups04_16",
+    "setregid01_16",
+    "setregid02_16",
+    "setregid03_16",
+    "setregid04_16",
+    "setresgid01_16",
+    "setresgid02_16",
+    "setresgid03_16",
+    "setresgid04_16",
+    "setresuid01_16",
+    "setresuid02_16",
+    "setresuid03_16",
+    "setresuid04_16",
+    "setresuid05_16",
+    "setreuid01_16",
+    "setreuid02_16",
+    "setreuid03_16",
+    "setreuid04_16",
+    "setreuid05_16",
+    "setreuid06_16",
+];
+
+const POSIX_MQ_SYSV_MSG_SEM_TASKS: [&str; 20] = [
+    "setreuid07_16",
+    "setuid01_16",
+    "setuid03_16",
+    "setuid04_16",
+    "mq_notify01",
+    "mq_notify02",
+    "mq_notify03",
+    "mq_open01",
+    "mq_timedreceive01",
+    "mq_timedsend01",
+    "mq_unlink01",
+    "msgctl12",
+    "msgget05",
+    "msgrcv05",
+    "msgrcv06",
+    "msgrcv07",
+    "msgrcv08",
+    "msgsnd06",
+    "semctl09",
+    "semget02",
+];
+
+const SYSV_SHM_CORE_TASKS: [&str; 19] = [
+    "shmat02", "shmat03", "shmat04", "shmat1", "shmctl01", "shmctl02", "shmctl03", "shmctl04",
+    "shmctl05", "shmctl06", "shmctl07", "shmctl08", "shmdt01", "shmdt02", "shmget03", "shmget04",
+    "shmget05", "shmget06", "shmt02",
+];
+
+// SysV SHM follow-up cases (passed on both musl and glibc).
+pub const SYSV_SHM_FOLLOWUP_TASKS: [&str; 9] = [
+    "shmget02", "shmt03", "shmt04", "shmt05", "shmt06", "shmt07", "shmt08", "shmt09", "shmt10",
+];
+
+// IPC namespace communication + POSIX MQ / SysV IPC namespace isolation.
+pub const IPC_NAMESPACE_TASKS: [&str; 12] = [
+    "msg_comm",
+    "sem_comm",
+    "shm_comm",
+    "shmem_2nstest",
+    "shmnstest",
+    "mesgq_nstest",
+    "mqns_01",
+    "mqns_02",
+    "mqns_03",
+    "mqns_04",
+    "sem_nstest",
+    "semtest_2ns",
+];
+
+// SysV message-queue stress case (passed on both musl and glibc).
+pub const SYSV_MSG_STRESS_TASKS: [&str; 1] = ["msgstress01"];
+
+// AIO / direct-I/O group.
+pub const AIO_DIO_CORE_TASKS: [&str; 34] = [
+    "aio-stress",
+    "aio01",
+    "aio02",
+    "aiocp",
+    "aiodio_append",
+    "aiodio_sparse",
+    "io_setup01",
+    "io_setup02",
+    "io_submit01",
+    "io_submit02",
+    "io_submit03",
+    "io_getevents01",
+    "io_getevents02",
+    "io_destroy01",
+    "io_destroy02",
+    "io_cancel01",
+    "io_cancel02",
+    "io_control01",
+    "io_pgetevents01",
+    "io_pgetevents02",
+    "readahead01",
+    "readahead02",
+    "dio_append",
+    "dio_read",
+    "dio_sparse",
+    "dio_truncate",
+    "diotest1",
+    "diotest2",
+    "diotest3",
+    "diotest4",
+    "diotest5",
+    "diotest6",
+    "doio",
+    "dma_thread_diotest",
+];
+
+// Memory-management group (mmap/mlock/madvise/memfd_create).
+pub const MM_MMAP_MADVISE_TASKS: [&str; 48] = [
+    "mmap-corruption01",
+    "mmap001",
+    "mmap08",
+    "mmap1",
+    "mmap2",
+    "mmap3",
+    "mmap10",
+    "mmap11",
+    "mmap12",
+    "mmap13",
+    "mmap14",
+    "mmap15",
+    "mmap16",
+    "mmap17",
+    "mmap18",
+    "mmap19",
+    "mmap20",
+    "mmapstress01",
+    "mmapstress02",
+    "mmapstress03",
+    "mmapstress04",
+    "mmapstress05",
+    "mmapstress06 20",
+    "mmapstress07 /tmp/mmapstress07_ltp.tmp",
+    "mmapstress08",
+    "mmapstress09 -p 20 -t 0.2",
+    "mmapstress10 -p 20 -t 0.2",
+    "mlock04",
+    "mlock05",
+    "mlock201",
+    "mlock202",
+    "mlock203",
+    "mlockall03",
+    "madvise03",
+    "madvise05",
+    "madvise06",
+    "madvise07",
+    "madvise08",
+    "madvise09",
+    "madvise10",
+    "madvise11",
+    "mincore04",
+    "page01",
+    "page02",
+    "memfd_create01",
+    "memfd_create02",
+    "memfd_create03",
+    "memfd_create04",
+];
+
+pub const MM_OOM_TASKS: [&str; 6] = [
+    "oom01",
+    "oom02",
+    "oom03",
+    "oom04",
+    "oom05",
+    "overcommit_memory",
+];
+
+// Threading / ptrace group.
+pub const THREADING_PTRACE_TASKS: [&str; 17] = [
+    "set_thread_area01",
+    "nptl01",
+    "ptrace01",
+    "ptrace02",
+    "ptrace03",
+    "ptrace04",
+    "ptrace05",
+    "ptrace06",
+    "ptrace07",
+    "ptrace08",
+    "ptrace09",
+    "ptrace10",
+    "ptrace11",
+    "pth_str01",
+    "pth_str02",
+    "pth_str03",
+    "run_sched_cliserv.sh",
+];
+
+// FS metadata / inotify / xattr / pidfd / prctl / ioctl / io_uring / namespace group.
+pub const FS_META_CHOWN_XATTR_TASKS: [&str; 30] = [
+    "chown01_16",
+    "chown02_16",
+    "chown03_16",
+    "chown04_16",
+    "chown05_16",
+    "fchown01_16",
+    "fchown02_16",
+    "fchown03_16",
+    "fchown04_16",
+    "fchown05_16",
+    "lchown01",
+    "lchown02",
+    "lchown03",
+    "lchown01_16",
+    "lchown02_16",
+    "lchown03_16",
+    "fgetxattr01",
+    "fgetxattr02",
+    "fgetxattr03",
+    "fsetxattr01",
+    "fsetxattr02",
+    "flistxattr01",
+    "flistxattr02",
+    "flistxattr03",
+    "fremovexattr01",
+    "fremovexattr02",
+    "llistxattr01",
+    "llistxattr02",
+    "llistxattr03",
+    "lremovexattr01",
+];
+
+pub const FS_META_INOTIFY_XATTR_TASKS: [&str; 44] = [
+    "chown01_16",
+    "chown02_16",
+    "chown03_16",
+    "chown04_16",
+    "chown05_16",
+    "fchown01_16",
+    "fchown02_16",
+    "fchown03_16",
+    "fchown04_16",
+    "fchown05_16",
+    "lchown01",
+    "lchown02",
+    "lchown03",
+    "lchown01_16",
+    "lchown02_16",
+    "lchown03_16",
+    "inotify01",
+    "inotify02",
+    "inotify03",
+    "inotify04",
+    "inotify05",
+    "inotify06",
+    "inotify07",
+    "inotify08",
+    "inotify09",
+    "inotify10",
+    "inotify11",
+    "inotify12",
+    "inotify_init1_01",
+    "inotify_init1_02",
+    "fgetxattr01",
+    "fgetxattr02",
+    "fgetxattr03",
+    "fsetxattr01",
+    "fsetxattr02",
+    "flistxattr01",
+    "flistxattr02",
+    "flistxattr03",
+    "fremovexattr01",
+    "fremovexattr02",
+    "llistxattr01",
+    "llistxattr02",
+    "llistxattr03",
+    "lremovexattr01",
+];
+
+pub const PIDFD_PRCTL_TASKS: [&str; 19] = [
+    "pidfd_getfd01",
+    "pidfd_getfd02",
+    "pidfd_open01",
+    "pidfd_open02",
+    "pidfd_open03",
+    "pidfd_open04",
+    "pidfd_send_signal01",
+    "pidfd_send_signal02",
+    "pidfd_send_signal03",
+    "prctl01",
+    "prctl02",
+    "prctl03",
+    "prctl04",
+    "prctl05",
+    "prctl06",
+    "prctl07",
+    "prctl08",
+    "prctl09",
+    "prctl10",
+];
+
+pub const IOCTL_IOURING_OPEN_TASKS: [&str; 26] = [
+    "ioctl01",
+    "ioctl02 -d /dev/tty",
+    "ioctl03",
+    "ioctl04",
+    "ioctl05",
+    "ioctl06",
+    "ioctl07",
+    "ioctl08",
+    "ioctl09",
+    "ioctl_loop01",
+    "ioctl_loop02",
+    "ioctl_loop03",
+    "ioctl_loop04",
+    "ioctl_loop05",
+    "ioctl_loop06",
+    "ioctl_loop07",
+    "ioctl_sg01",
+    "io_uring01",
+    "io_uring02",
+    "open12",
+    "open14",
+    "openat02",
+    "openat04",
+    "openat201",
+    "openat202",
+    "openat203",
+];
+
+pub const NS_MOUNT_CORE_TASKS: [&str; 10] = [
+    "mountns01",
+    "mountns02",
+    "mountns03",
+    "mountns04",
+    "setns01",
+    "setns02",
+    "unshare01",
+    "unshare02",
+    "timens01",
+    "pidns01",
+];
+
+pub const KCMP_TASKS: [&str; 1] = ["kcmp01"];
+
+// Fanotify / mount-API / namespace follow-up / pidns / module /
+// misc device-query (ioprio/iopl/perf_event/sysinfo/pathconf) group.
+pub const FANOTIFY_CORE_TASKS: [&str; 20] = [
+    "fanotify01",
+    "fanotify02",
+    "fanotify03",
+    "fanotify04",
+    "fanotify05",
+    "fanotify06",
+    "fanotify07",
+    "fanotify08",
+    "fanotify09",
+    "fanotify10",
+    "fanotify11",
+    "fanotify12",
+    "fanotify13",
+    "fanotify14",
+    "fanotify15",
+    "fanotify16",
+    "fanotify17",
+    "fanotify18",
+    "fanotify19",
+    "fanotify20",
+];
+
+pub const MOUNT_API_TASKS: [&str; 20] = [
+    "fanotify21",
+    "fanotify22",
+    "fanotify23",
+    "mount01",
+    "mount02",
+    "mount03",
+    "mount04",
+    "mount05",
+    "mount06",
+    "mount07",
+    "umount01",
+    "umount02",
+    "umount03",
+    "umount2_01",
+    "umount2_02",
+    "mount_setattr01",
+    "fsconfig01",
+    "fsconfig02",
+    "fsconfig03",
+    "fsopen01",
+];
+
+pub const NS_MOUNT_FOLLOWUP_TASKS: [&str; 20] = [
+    "fsopen02",
+    "fsmount01",
+    "fsmount02",
+    "fspick01",
+    "fspick02",
+    "open_tree01",
+    "open_tree02",
+    "move_mount01",
+    "move_mount02",
+    "userns01",
+    "userns02",
+    "userns03",
+    "userns04",
+    "userns05",
+    "userns06",
+    "userns07",
+    "userns08",
+    "pidns02",
+    "pidns03",
+    "pidns04",
+];
+
+pub const PIDNS_MODULE_TASKS: [&str; 20] = [
+    "pidns05",
+    "pidns06",
+    "pidns10",
+    "pidns12",
+    "pidns13",
+    "pidns16",
+    "pidns17",
+    "pidns20",
+    "pidns30",
+    "pidns31",
+    "pidns32",
+    "init_module01",
+    "init_module02",
+    "finit_module01",
+    "finit_module02",
+    "delete_module01",
+    "delete_module02",
+    "delete_module03",
+    "getcpu01",
+    "membarrier01",
+];
+
+// `pathconf02` stays in this batch, but submit_plan currently runs it only
+// in glibc lanes because musl's wrapper does not validate path errors here.
+pub const IO_PERF_SYSINFO_PATHCONF_TASKS: [&str; 20] = [
+    "ioprio_get01",
+    "ioprio_set01",
+    "ioprio_set02",
+    "ioprio_set03",
+    "ioperm01",
+    "ioperm02",
+    "iopl01",
+    "iopl02",
+    "perf_event_open01",
+    "perf_event_open02",
+    "perf_event_open03",
+    "sysinfo01",
+    "sysinfo02",
+    "sysinfo03",
+    "personality01",
+    "personality02",
+    "confstr01",
+    "pathconf01",
+    "pathconf02",
+    "fpathconf01",
+];
+
+// Cgroup / memcg / security (keys, crypto, BPF, CVE) /
+// uname-sysfs-aslr misc-features group.
+pub const CGROUP_CORE_CPU_TASKS: [&str; 20] = [
+    "cgroup_core01",
+    "cgroup_core02",
+    "cgroup_core03",
+    "cgroup_xattr",
+    "cfs_bandwidth01",
+    "cgroup_regression_test.sh",
+    "cgroup_fj_function.sh debug",
+    "cgroup_fj_function.sh cpuset",
+    "cgroup_fj_function.sh cpu",
+    "cgroup_fj_function.sh cpuacct",
+    "cgroup_fj_function.sh memory",
+    "cgroup_fj_function.sh freezer",
+    "cgroup_fj_function.sh devices",
+    "cgroup_fj_function.sh blkio",
+    "cgroup_fj_function.sh net_cls",
+    "cgroup_fj_function.sh perf_event",
+    "cgroup_fj_function.sh net_prio",
+    "cgroup_fj_function.sh hugetlb",
+    "cpuacct.sh 1 1",
+    "cpuset01",
+];
+
+pub const CGROUP_MEM_PID_TASKS: [&str; 28] = [
+    "pids.sh 1 5 0",
+    "pids.sh 2 5 0",
+    "pids.sh 3 5 0",
+    "pids.sh 5 5 0",
+    "pids.sh 6 5 0",
+    "pids.sh 7 6 3",
+    "pids.sh 8 5 0",
+    "pids.sh 9 5 0",
+    "memcg_control_test.sh",
+    "memcg_failcnt.sh",
+    "memcg_force_empty.sh",
+    "memcg_limit_in_bytes.sh",
+    "memcg_max_usage_in_bytes_test.sh",
+    "memcg_memsw_limit_in_bytes_test.sh",
+    "memcg_move_charge_at_immigrate_test.sh",
+    "memcg_regression_test.sh",
+    "memcg_stat_rss.sh",
+    "memcg_stat_test.sh",
+    "memcg_stress_test.sh",
+    "memcg_subgroup_charge.sh",
+    "memcg_usage_in_bytes_test.sh",
+    "memcg_use_hierarchy_test.sh",
+    "run_memctl_test.sh 1",
+    "run_memctl_test.sh 2",
+    "memcontrol01",
+    "memcontrol02",
+    "memcontrol03",
+    "memcontrol04",
+];
+
+pub const SECURITY_KEYS_CRYPTO_TASKS: [&str; 20] = [
+    "keyctl01",
+    "keyctl02",
+    "keyctl03",
+    "keyctl04",
+    "keyctl05",
+    "keyctl06",
+    "keyctl07",
+    "keyctl08",
+    "keyctl09",
+    "request_key01",
+    "request_key02",
+    "request_key03",
+    "request_key04",
+    "request_key05",
+    "af_alg01",
+    "af_alg02",
+    "af_alg03",
+    "af_alg04",
+    "af_alg05",
+    "af_alg06",
+];
+
+pub const SECURITY_BPF_CVE_TASKS: [&str; 20] = [
+    "af_alg07",
+    "crypto_user01",
+    "crypto_user02",
+    "pcrypt_aead01",
+    "bpf_map01",
+    "bpf_prog01",
+    "bpf_prog02",
+    "bpf_prog03",
+    "bpf_prog04",
+    "bpf_prog05",
+    "bpf_prog06",
+    "bpf_prog07",
+    "cve-2014-0196",
+    "cve-2015-3290",
+    "cve-2016-10044",
+    "cve-2016-7042",
+    "cve-2016-7117",
+    "cve-2017-16939",
+    "cve-2017-17052",
+    "cve-2017-17053",
+];
+
+pub const UNAME_SYSFS_ASLR_TASKS: [&str; 20] = [
+    "cve-2017-2618",
+    "cve-2017-2671",
+    "cve-2022-4378",
+    "newuname01",
+    "utsname01",
+    "utsname02",
+    "utsname03",
+    "utsname04",
+    "sysconf01",
+    "getpagesize01",
+    "syscall01",
+    "sysfs01",
+    "sysfs02",
+    "sysfs03",
+    "sysfs04",
+    "sysfs05",
+    "sysctl01",
+    "sysctl03",
+    "sysctl04",
+    "aslr01",
+];
+
+// High-priority core-100 follow-up across process / signal / cred /
+// IPC / FD / time / file-metadata. Excludes stubbed interfaces such
+// as timerfd / clone3 and sysctl-driven SysV IPC capacity controls.
+pub const CORE100_TASKS: [&str; 100] = [
+    "clone05",
+    "clone06",
+    "clone07",
+    "wait403",
+    "execve05",
+    "execve06",
+    "execveat03",
+    "sigaction01",
+    "sigaction02",
+    "rt_sigaction01",
+    "rt_sigaction02",
+    "rt_sigaction03",
+    "sigaltstack01",
+    "sigaltstack02",
+    "sigpending02",
+    "sigprocmask01",
+    "sigrelse01",
+    "sigsuspend01",
+    "sigtimedwait01",
+    "sigwait01",
+    "sigwaitinfo01",
+    "signal01",
+    "signal02",
+    "signal03",
+    "signal04",
+    "signal05",
+    "chroot01",
+    "chroot02",
+    "chroot03",
+    "chroot04",
+    "getgroups01",
+    "setgroups01",
+    "setgroups02",
+    "msgctl03",
+    "msgctl04",
+    "msgctl05",
+    "msgsnd02",
+    "msgsnd05",
+    "msgrcv02",
+    "msgrcv03",
+    "semctl03",
+    "semctl04",
+    "semctl05",
+    "semctl07",
+    "semctl08",
+    "semop03",
+    "dup01",
+    "dup02",
+    "dup03",
+    "dup04",
+    "dup05",
+    "dup06",
+    "dup07",
+    "dup201",
+    "dup202",
+    "dup203",
+    "dup204",
+    "dup205",
+    "dup206",
+    "dup207",
+    "dup3_01",
+    "dup3_02",
+    "umask01",
+    "copy_file_range01",
+    "copy_file_range02",
+    "copy_file_range03",
+    "clock_gettime01",
+    "clock_gettime02",
+    "clock_gettime03",
+    "clock_gettime04",
+    "clock_settime01",
+    "clock_settime02",
+    "clock_settime03",
+    "clock_getres01",
+    "clock_nanosleep01",
+    "clock_nanosleep02",
+    "clock_nanosleep03",
+    "clock_nanosleep04",
+    "nanosleep01",
+    "nanosleep02",
+    "nanosleep04",
+    "gettimeofday02",
+    "fchmod01",
+    "fchmod02",
+    "fchmod03",
+    "fchmod04",
+    "fchmod05",
+    "fchmod06",
+    "fchmodat01",
+    "fchmodat02",
+    "stat01",
+    "stat02",
+    "stat03",
+    "lstat01",
+    "lstat02",
+    "fstat02",
+    "fstat03",
+    "fstatat01",
+    "statfs02",
+    "statfs03",
+];
