+++ user/src/bin/ltp_dependence/tasks/smoke.rs
@@ -1,0 +1,11 @@
+//! Smoke / sanity sample: a small reachability probe.
+//
+// Auto-extracted from the original flat `mod.rs`; test names are
+// preserved verbatim so the set consumed by `submit_plan.rs` is unchanged.
+
+#![allow(dead_code)]
+
+pub const LTP_TEST_POINTS: [&str; 10] = [
+    "abort01", "abs01", "accept01", "access01", "access02", "access03", "access04", "acct01",
+    "acct02", "write01",
+];
